frame: PAYLOAD_TYPE -> FRAME_TYPE and move to frame API
Command frames sent to bootloader do not have 0x30 before payload. This
makes the 0x30 appear as frame type, rather than a fixed payload header.
It needs to be escaped and checksummed, though.

This commit adds frame types for bootloader, and add a parameter for
setting this in frame_send(). Thus we can move the 0x30 from requests.c
to frame.c and make it common between MCU app and bootloader.

While at it, fix some issue:

1. Check for failed malloc's by == NULL, not <= 0.
2. Clear RX buffer after delivering a frame.